% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anforandetext_replace_collocations.R
\name{anforandetext_replace_collocation}
\alias{anforandetext_replace_collocation}
\title{Replace collocations in anforandetext}
\usage{
anforandetext_replace_collocation(anforandetext, collocation_folder)
}
\arguments{
\item{anforandetext}{a character vector from anforandetext in corpus file}

\item{collocation_folder}{a path to a folder with files with one collocation per row.}
}
\description{
Replace collocations in data-raw collocations
}
\examples{
collocation_folder <- "RPackage/data-raw/collocations/"
anforandetext_replace_collocation(txt, collocation_folder)

}
\keyword{Internal}

