% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anforandetext_replace_dash_tokens.R
\name{anforandetext_replace_dash_tokens}
\alias{anforandetext_replace_dash_tokens}
\title{Replace dashed tokens in anforandetext}
\usage{
anforandetext_replace_dash_tokens(speeches)
}
\arguments{
\item{speeches}{a speeches corpus object}
}
\description{
Replace dashed tokens in anforandetext
}
\details{
This function handles dashes used in the corpus. 
There are two main uses for dashes: 
1) New lines that cuts words and 
2) Combined words
}
\keyword{Internal}
