% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_collocation_size}
\alias{compute_collocation_size}
\title{Compute the vocabulary size of a corpus}
\usage{
compute_collocation_size(collocation_folder)
}
\arguments{
\item{collocation_folder}{a path to a folder with files with one collocation per row.}
}
\description{
Compute the vocabulary size of a corpus
}
\keyword{Internal}
