% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_vocabulary_size}
\alias{compute_vocabulary_size}
\title{Compute the vocabulary size of a corpus}
\usage{
compute_vocabulary_size(corpus)
}
\arguments{
\item{corpus}{a \code{speeches} corpus to compute vocabulary size for.}
}
\description{
Compute the vocabulary size of a corpus
}
\keyword{Internal}

