% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_debate_ids.R
\name{create_debate_ids}
\alias{create_debate_ids}
\title{Compute a test_set indicator}
\usage{
create_debate_ids(corpus)
}
\arguments{
\item{corpus}{a \code{speeches} corpus.}
}
\description{
Compute a test_set indicator
}
\keyword{Internal}
