% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_debate_type.R
\name{create_debate_type}
\alias{create_debate_type}
\title{Create debate type}
\usage{
create_debate_type(avsnittsrubrik, kammaraktivitet)
}
\arguments{
\item{avsnittsrubrik}{Avsnittsrubrik to use to identify \code{debate_type}}

\item{kammaraktivitet}{Kammaraktivitet to use to identify \code{debate_type}}
}
\description{
Create debate type
}
\keyword{Internal}
