% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_role.R
\name{create_role}
\alias{create_role}
\title{Clean the parti variable in the corpus file}
\usage{
create_role(talare)
}
\arguments{
\item{talare}{the raw talare variable}
}
\description{
Clean the parti variable in the corpus file
}
\keyword{Internal}

