% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_set.R
\name{create_test_set}
\alias{create_test_set}
\title{Compute a test_set indicator}
\usage{
create_test_set(corpus, seed = 4711, size = 10000)
}
\arguments{
\item{corpus}{a \code{speeches} corpus.}

\item{seed}{the \code{seed} to generate create the test set for.}

\item{size}{the size of the test set.}
}
\description{
Compute a test_set indicator
}
\keyword{Internal}
