% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parti_clean.R
\name{parti_clean}
\alias{parti_clean}
\title{Clean the parti variable in the corpus file}
\usage{
parti_clean(parti)
}
\arguments{
\item{parti}{a character vector from anforandetext in corpus file}
}
\description{
Clean the parti variable in the corpus file
}
\keyword{Internal}

