% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrpsriksdag.R
\docType{package}
\name{rcrpsriksdag-package}
\alias{rcrpsriksdag-package}
\alias{rcrpsriksdag}
\title{rcrpsriksdag R package}
\description{
An R package that contain corpora from the Swedish parliament (Riksdagen).
}
