% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_anforande_json_zip.R
\name{read_anforande_json_zip}
\alias{parse_anforande_json}
\alias{read_anforande_json_zip}
\title{Read in a json-zip file}
\usage{
read_anforande_json_zip(anforande_json_zip)

parse_anforande_json(json_file)
}
\arguments{
\item{anforande_json_zip}{Anforande json zip file}

\item{json_file}{a json file from an anforande json zip file.}
}
\value{
a \code{tbl_df} \code{data.frame}
}
\description{
Read in a json-zip file
}
\examples{
anforande_json_zip <- "RPackage/data-raw/raw_data/anforande-201516.json.zip" 
res <- read_anforande_json_zip(anforande_json_zip)

}
\keyword{Internal}

