% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anforandetext_replace_tokens.R
\name{read_in_token_error_files}
\alias{read_in_token_error_files}
\title{Read in the files in collocation folders}
\usage{
read_in_token_error_files(token_errors_folder)
}
\arguments{
\item{collocation_folder}{a path to a folder with files with one collocation per row.}
}
\description{
Read in the files in collocation folders
}
\keyword{Internal}

