% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_anforande_folder.R
\name{read_json_anforande_folder}
\alias{read_json_anforande_folder}
\title{Read in all json-zip files in a folder}
\usage{
read_json_anforande_folder(folder)
}
\arguments{
\item{folder}{A folder with anforande-NNNNNNNN.json.zip files.}
}
\value{
a \code{tbl_df} \code{data.frame}
}
\description{
Read in all json-zip files in a folder
}
\examples{
folder <- "RPackage/data-raw/raw_data/"
corpus <- read_anforande_json_zip(folder)

}
\keyword{Internal}

