% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pattern_context.R
\name{show_pattern_context}
\alias{show_pattern_context}
\title{Show a pattern within a textual context.}
\usage{
show_pattern_context(txt, pattern, character_window = 30)
}
\arguments{
\item{txt}{A character vector to find the pattern within}

\item{pattern}{The pattern to show the context for.}

\item{character_window}{The context size in number of characters.}
}
\description{
Show a pattern within a textual context.
}
\keyword{Internal}
