% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{speeches}
\alias{speeches}
\title{Debates in the Swedish parliament (cleaned corpus)}
\format{A \code{tbl_df} with one row per speech.}
\usage{
speeches
}
\description{
A corpus with one speech per row. 
This is a cleaned version of \code{speeches_raw}, see  
RPackage/data-raw/speeches.R on how the data has been cleaned.
}
\seealso{
\link{speeches_raw}
}
\keyword{datasets}

