% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speeches_preprocessing.R
\name{speeches_parse_variables}
\alias{speeches_parse_variables}
\alias{speeches_remove_observations}
\title{Preprocessing of the speeches corpus}
\usage{
speeches_parse_variables(corpus)

speeches_remove_observations(corpus)
}
\arguments{
\item{corpus}{The speeches corpus.}
}
\description{
Parse the variables into correct R format.
}
\keyword{Internal}

