% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{speeches_raw}
\alias{speeches_raw}
\title{Debates in the Swedish parliament (raw corpus)}
\format{A \code{tbl_df} with one row debate speech.}
\usage{
speeches_raw
}
\description{
Raw data containing debates in the Swedish parliament.
}
\seealso{
\link{speeches}
}
\keyword{datasets}

