% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{speeches_sentence}
\alias{speeches_sentence}
\title{Debates in the Swedish parliament (cleaned corpus)}
\format{A \code{tbl_df} with one row debate speech.}
\description{
Debates in the Swedish parliament (cleaned corpus)
}
\details{
A corpus with one sentence per row. 
The debates has been cut into sentences using the Gavagai API.
This is a cleaned version of \code{speeches_sentence_raw}, see  
\code{RPackage/data-raw/gavagai.R} on how the data has been cleaned.
}
\seealso{
\link{speeches_sentence_raw} \link{speeches}
}
\keyword{data}
