% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{speeches_sentence_raw}
\alias{speeches_sentence_raw}
\title{Debates by sentence in the Swedish parliament (raw corpus)}
\format{A \code{tbl_df} with one row per sentence in debate speeches.}
\description{
Debates by sentence in the Swedish parliament (raw corpus)
}
\details{
Raw data containing debates in the Swedish parliament. 
The debates has been cut into sentences using the Gavagai API.
}
\seealso{
\link{speeches_sentence} \link{speeches_raw}
}
\keyword{data}
