% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_pattern_tokens.R
\name{tabulate_pattern_tokens}
\alias{tabulate_pattern_tokens}
\title{Tablulate tokens identified by pattern}
\usage{
tabulate_pattern_tokens(txt, pattern)
}
\arguments{
\item{txt}{A character vector to find the pattern within}

\item{pattern}{The pattern to extract tokens from.}
}
\description{
Tablulate tokens identified by pattern
}
\keyword{Internal}

