% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talare_clean.R
\name{talare_clean}
\alias{talare_clean}
\title{Clean the parti variable in the corpus file}
\usage{
talare_clean(talare)
}
\arguments{
\item{parti}{a character vector from anforandetext in corpus file}
}
\description{
Clean the parti variable in the corpus file
}
\keyword{Internal}

