% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anforandetext_replace_tokens.R
\name{anforandetext_replace_token_errors}
\alias{anforandetext_replace_token_errors}
\title{Replace token errors in anforandetext}
\usage{
anforandetext_replace_token_errors(speeches, token_errors_folder)
}
\arguments{
\item{token_errors_folder}{a path to a folder with files with one collocation per row.}

\item{anforandetext}{a character vector from anforandetext in corpus file}
}
\description{
Replace token errors in anforandetext
}
\examples{
token_errors_folder <- "RPackage/data-raw/dash_errors"
token_errors_folder <- "RPackage/data-raw/token_errors/"
anforandetext_replace_collocation(txt, token_errors_folder)

}
\keyword{Internal}
