% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parti_clean.R
\name{parti_clean}
\alias{parti_clean}
\title{Clean the parti variable in the corpus file}
\usage{
parti_clean(parti, talare,
  party_errors_file = "RPackage/data-raw/data_errors/party_errors.csv")
}
\arguments{
\item{parti}{a factor vector from parti in a \code{speeches} \code{data_frame}}

\item{talare}{a factor vector from talare in a \code{speeches} \code{data_frame}}

\item{party_errors_file}{a csv file containing the variables \code{wrong_parti}, \code{true_parti} and \code{talare}}
}
\description{
Clean the parti variable in the corpus file
}
\keyword{Internal}
