% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{speeches_raw}
\alias{speeches_raw}
\title{Debates in the Swedish parliament (raw corpus)}
\format{A \code{tbl_df} with one row debate speech.}
\description{
Debates in the Swedish parliament (raw corpus)
}
\details{
Raw data containing debates in the Swedish parliament.
}
\seealso{
\link{speeches} \link{speeches_sentence_raw}
}
\keyword{data}
