% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/stop_words.R
\docType{data}
\name{stop_words_se}
\alias{stop_words_se}
\alias{stop_words_se}
\title{Stop word lists}
\format{An object of class \code{character} of length 116.}
\usage{
stop_words_se
}
\description{
Stop word lists

Stop word lists
}
\details{
A stopword list of common Swedish stop words.
}
\keyword{data}
\keyword{datasets}
