% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talare_clean.R
\name{talare_clean}
\alias{talare_clean}
\alias{get_speaker_errors}
\title{Clean the parti variable in the corpus file}
\usage{
talare_clean(talare)

get_speaker_errors()
}
\arguments{
\item{talare}{a character vector from speaker in corpus file}
}
\description{
Clean the parti variable in the corpus file

Get errors in speaker from csv file
}
\keyword{Internal}
